Fix dirty overload of Keccak function
The Keccak function was overloaded with a set of options to generate
JSON files. This commit separates this logic into `keccak_gen`, which
hopefully can also be adopted as a pattern for upcoming work on
`sm/sha256` and `sm/blake`.
