+++ tools/sm/keccak_f/keccak.cpp
@@ -6,7 +6,7 @@
     Input is a buffer of any length, including 0
     Output is 256 bits long buffer containing the 32B keccak hash of the input
 */
-void Keccak (const uint8_t * pInput, uint64_t inputSize, uint8_t * pOutput, string scriptFile, string polsFile, string connectionsFile)
+void Keccak (const uint8_t * pInput, uint64_t inputSize, uint8_t * pOutput)
 {
     KeccakInput input;
     input.init(pInput, inputSize);
@@ -21,51 +21,10 @@ void Keccak (const uint8_t * pInput, uint64_t inputSize, uint8_t * pOutput, stri
         }
         KeccakF(S);
         S.printCounters();
-
-        // Generate the script file only after the first keccak-f round
-        if (scriptFile.size() > 0)
-        {
-            json j;
-            S.saveScriptToJson(j);
-            cout << "Generating Keccak script file: " << scriptFile << endl;
-            json2file(j, scriptFile);
-            cout << "Generated Keccak script file: " << scriptFile << endl;
-            scriptFile = "";
-        }
-
-        // Generate the polynomials file only after the first keccak-f round
-        if (polsFile.size() > 0)
-        {
-            json j;
-            S.savePolsToJson(j);
-            cout << "Generating Keccak polynomials file: " << polsFile << endl;
-            json2file(j, polsFile);
-            cout << "Generated Keccak polynomials file: " << polsFile << endl;
-            polsFile = "";
-        }
-
-        // Generate the connections file only after the first keccak-f round
-        if (connectionsFile.size() > 0)
-        {
-            json j;
-            S.saveConnectionsToJson(j);
-            cout << "Generating Keccak connections file: " << connectionsFile << endl;
-            json2file(j, connectionsFile);
-            cout << "Generated Keccak connections file: " << connectionsFile << endl;
-            connectionsFile = "";
-        }
         
         S.copySoutToSinAndResetRefs();
     }
     S.getOutput(pOutput);
-}
-
-void KeccakGenerateScript (const Config & config)
-{
-    TimerStart(KECCAK_SM_GENERATE_SCRIPT);
-    uint8_t hash[32];
-    Keccak(NULL, 0, hash, config.keccakScriptFile, config.keccakPolsFile, config.keccakConnectionsFile);
-    TimerStopAndLog(KECCAK_SM_GENERATE_SCRIPT);
 }
 
 /* Unit test */
@@ -90,7 +49,7 @@ void KeccakSMTest (void)
         0x60, 0x51, 0x0B, 0x48, 0xC6, 0x01, 0x27, 0x42, 0xC5, 0x40, 
         0xA8, 0x1A, 0xBA, 0x4B, 0xCA, 0x2F, 0x78, 0xB9, 0xD1, 0x4B, 
         0xFD, 0x2F, 0x12, 0x3E, 0x2E, 0x53, 0xEA, 0x3E, 0x61, 0x7B, 
-        0x3A, 0x35, 0x28, 0xF9, 0xCD, 0xD6,   0x63, 0x0F, 0xD3, 0x30, 
+        0x3A, 0x35, 0x28, 0xF9, 0xCD, 0xD6, 0x63, 0x0F, 0xD3, 0x30, 
         0x1B, 0x9C, 0x89, 0x11, 0xF7, 0xBF, 0x06, 0x3D, 0x29, 0x90,
 
         0x27, 0xCC, 0x1E, 0xE6, 0x56, 0x7E, 0x0F, 0xE5, 0xD6, 0x64,
